Fix clean (avoid '!')

+++ .bashrc
@@ -152,8 +152,11 @@ alias rmr='rm -r'
 alias cd..='cd ..'
 
 function clean() {
-    find ${1:-./} -name "*~" -print -delete 2>/dev/null
-    find ${1:-./} -name ".*.swp" -print -delete 2>/dev/null
+    if [[ $1 != '!' ]]
+    then
+        find ${1:-./} -name "*~" -print -delete 2>/dev/null
+        find ${1:-./} -name ".*.swp" -print -delete 2>/dev/null
+    fi
 }
 
 function psg() {
